requirement.txt
attrs==20.2.0
Django==2.2
django-cors-headers==0.1
django-filter==2.4.0
django-model-utils==4.0.0
django-money==1.1
djangorestframework==3.12.1
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
py-moneyed==0.8.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
zipp==3.4.0
